improve pylint
