+++ pypowsybl/flowdecomposition/impl/parameters.py
