+++ pypowsybl/flowdecomposition/impl/util.py
