+++ pypowsybl/glsk.py
@@ -11,7 +11,7 @@ from typing import (
 from os import PathLike as _PathLike
 from datetime import datetime
 from pypowsybl import _pypowsybl
-from pypowsybl.network import Network as _Network, _path_to_str
+from pypowsybl.network.impl.network import Network as _Network, _path_to_str
 
 
 class GLSKDocument:
