+++ pypowsybl/loadflow/impl/component_result.py
