+++ pypowsybl/network/impl/bus_breaker_topology.py
@@ -4,11 +4,11 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 # SPDX-License-Identifier: MPL-2.0
 #
-import pypowsybl._pypowsybl as _pp
 import networkx as nx
 from pandas import DataFrame
+import pypowsybl._pypowsybl as _pp
 
-from pypowsybl.network.impl.util import create_data_frame_from_series_array
+from pypowsybl.network.impl.util import _create_data_frame_from_series_array
 
 
 class BusBreakerTopology:
@@ -22,11 +22,11 @@ class BusBreakerTopology:
     """
 
     def __init__(self, network_handle: _pp.JavaHandle, voltage_level_id: str):
-        self._elements = create_data_frame_from_series_array(
+        self._elements = _create_data_frame_from_series_array(
             _pp.get_bus_breaker_view_elements(network_handle, voltage_level_id))
-        self._switchs = create_data_frame_from_series_array(
+        self._switchs = _create_data_frame_from_series_array(
             _pp.get_bus_breaker_view_switches(network_handle, voltage_level_id))
-        self._buses = create_data_frame_from_series_array(
+        self._buses = _create_data_frame_from_series_array(
             _pp.get_bus_breaker_view_buses(network_handle, voltage_level_id))
 
     @property
