+++ pypowsybl/network/impl/perunit.py
@@ -10,7 +10,8 @@ import pandas as pd
 import numpy as np
 from numpy.typing import ArrayLike
 from pypowsybl import _pypowsybl
-from pypowsybl.network.impl.network import ElementType, Network
+from pypowsybl.network.impl.network import ElementType
+from pypowsybl.network import Network
 from pypowsybl.utils.dataframes import _adapt_df_or_kwargs
 
 
@@ -487,7 +488,7 @@ class PerUnitView:  # pylint: disable=too-many-public-methods
         self._un_per_unit_v(to_update, ['rated_u1'], nominal_v1)
         self._un_per_unit_v(to_update, ['rated_u2'], nominal_v2)
         self._un_per_unit_v(to_update, ['rated_u3'], nominal_v3)
-        self._network._update_elements(_pypowsybl.ElementType.THREE_WINDINGS_TRANSFORMER, to_update)
+        self._network.update_3_windings_transformers(to_update)
 
     def update_lcc_converter_station(self, df: pd.DataFrame = None, **kwargs: ArrayLike) -> None:
         """
@@ -497,7 +498,7 @@ class PerUnitView:  # pylint: disable=too-many-public-methods
         nominal_v = self._get_indexed_nominal_v(self._network.get_lcc_converter_stations())
         self._un_per_unit_p(to_update, ['p', 'q'])
         self._un_per_unit_i(to_update, ['i'], nominal_v)
-        self._network._update_elements(_pypowsybl.ElementType.LCC_CONVERTER_STATION, to_update)
+        self._network.update_lcc_converter_stations(to_update)
 
 
 def per_unit_view(network: Network, sn: float = 100) -> PerUnitView:
