+++ pypowsybl/network/impl/svg.py
