+++ pypowsybl/security/impl/increased_violations_parameters.py
@@ -4,7 +4,7 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 # SPDX-License-Identifier: MPL-2.0
 #
-class IncreasedViolationsParameters:
+class IncreasedViolationsParameters:   # pylint: disable=too-few-public-methods
     """
     Parameters which define what violations should be considered as "increased" between N and post-contingency situations
