+++ pypowsybl/sensitivity/impl/ac_sensitivity_analysis.py
@@ -49,10 +49,8 @@ class AcSensitivityAnalysis(SensitivityAnalysis):
         """
         sensitivity_parameters = Parameters(load_flow_parameters=parameters) if isinstance(parameters,
                                                                                            LfParameters) else parameters
-        p: _pypowsybl.SensitivityAnalysisParameters = sensitivity_parameters._to_c_parameters() if sensitivity_parameters is not None else Parameters()._to_c_parameters()
+        p: _pypowsybl.SensitivityAnalysisParameters = sensitivity_parameters._to_c_parameters() if sensitivity_parameters is not None else Parameters()._to_c_parameters()  # pylint: disable=W0212
         return AcSensitivityAnalysisResult(
-            _pypowsybl.run_sensitivity_analysis(self._handle, network._handle, False, p, provider,
-                                                None if reporter is None else reporter._reporter_model),
-            # pylint: disable=protected-access
+            _pypowsybl.run_sensitivity_analysis(self._handle, network._handle, False, p, provider, None if reporter is None else reporter._reporter_model), # pylint: disable=protected-access
             branches_ids=self.branches_ids, branch_data_frame_index=self.branch_data_frame_index,
             bus_ids=self.bus_voltage_ids, target_voltage_ids=self.target_voltage_ids)
