+++ pypowsybl/sensitivity/impl/dc_sensitivity_analysis_result.py
@@ -78,4 +78,3 @@ class DcSensitivityAnalysisResult:
             return None
         data = np.array(matrix, copy=False)
         return pd.DataFrame(data=data, columns=self.branches_ids[matrix_id], index=['reference_flows'])
-
